handwrite/requirements.txt
pillow
python-docx
handright